refactor: split BackupManager files below 500-line limit (AMUX-230)
BackupManager.swift (580) and BackupManagerQueueIntegration.swift (602)
exceeded the workspace 500-line limit, making verify-code.sh warn on
every edit. Pure code movement into focused extension files:

- BackupManager+Forwarding.swift: modal flags, destination/source
  forwarding, progress + scanning delegation
- BackupManager+UITestSupport.swift: loadUITestPaths
- BackupManager+Preflight.swift: buildPreflightSummary
- BackupManager+Completion.swift: handleBackupCompletion
- BackupManager+LargeBackup.swift: checkForLargeBackupAndWait +
  respondToLargeBackupConfirmation

loadUITestPaths and buildPreflightSummary go private -> internal
(cross-file extensions can't see private); no behavioral changes.
Root files now 433 and 482 lines. 58/58 BackupManager +
PreferencesProviding tests green before and after.
